src/mini_apps/circuit_optimization/requirements.txt
qiskit==1.2.4
qiskit-aer==0.16.0
qiskit-aer-gpu==0.15.1
qiskit-addon-cutting==0.9.0
qiskit-ionq==0.5.12
qiskit-ibm-provider==0.11.0
qiskit-ibm-runtime==0.29.0
PennyLane==0.40.0
PennyLane-qiskit==0.40.1
PennyLane_Lightning==0.40.0
Pilot-Quantum==0.31.23
ray==2.34.0
numpy==2.0.2
matplotlib==3.10.1
scipy==1.15.2
scikit-learn==1.6.1
seaborn==0.13.2
pandas==2.2.3
networkx==3.4.2